requirements.txt
streamlit
fastapi
uvicorn
python-multipart
requests
torch
torchvision
pillow
numpy
